Abstracted `parsePartitionRow` to be defined in engine specific impls
