+++ hudi-spark-datasource/hudi-spark/src/main/scala/org/apache/hudi/AbstractHoodieTableFileIndex.scala
@@ -28,9 +28,7 @@ import org.apache.hudi.common.table.HoodieTableMetaClient
 import org.apache.hudi.common.table.view.{FileSystemViewStorageConfig, HoodieTableFileSystemView}
 import org.apache.spark.sql.execution.datasources.{FileStatusCache, NoopCache}
 import org.apache.spark.sql.hudi.HoodieSqlUtils
-import org.apache.spark.unsafe.types.UTF8String
 
-import java.nio.charset.StandardCharsets
 import scala.collection.JavaConversions._
 import scala.collection.JavaConverters._
 import scala.collection.mutable
@@ -251,67 +249,20 @@ abstract class AbstractHoodieTableFileIndex(engineContext: HoodieEngineContext,
 
     val partitionSchema = _partitionColumns
 
-    // Convert partition path to PartitionRowPath
+    // Convert partition's path into partition descriptor
     partitionPaths.map { partitionPath =>
-      val partitionRow: Array[Any] = if (partitionSchema.length == 0) {
-        // This is a non-partitioned table
-        Array.empty
-      } else {
-        val partitionFragments = partitionPath.split("/")
-
-        if (partitionFragments.length != partitionSchema.length &&
-          partitionSchema.length == 1) {
-          // If the partition column size is not equal to the partition fragment size
-          // and the partition column size is 1, we map the whole partition path
-          // to the partition column which can benefit from the partition prune.
-          val prefix = s"${partitionSchema.head}="
-          val partitionValue = if (partitionPath.startsWith(prefix)) {
-            // support hive style partition path
-            partitionPath.substring(prefix.length)
-          } else {
-            partitionPath
-          }
-          // TODO replace w/ byte array
-          Array(UTF8String.fromString(partitionValue))
-        } else if (partitionFragments.length != partitionSchema.length &&
-          partitionSchema.length > 1) {
-          // If the partition column size is not equal to the partition fragments size
-          // and the partition column size > 1, we do not know how to map the partition
-          // fragments to the partition columns. So we trait it as a Non-Partitioned Table
-          // for the query which do not benefit from the partition prune.
-          logWarning(s"Cannot do the partition prune for table $basePath." +
-            s"The partitionFragments size (${partitionFragments.mkString(",")})" +
-            s" is not equal to the partition columns size(${partitionSchema.mkString(",")})")
-          Array.empty
-        } else { // If partitionSeqs.length == partitionSchema.fields.length
-
-          // Append partition name to the partition value if the
-          // HIVE_STYLE_PARTITIONING is disable.
-          // e.g. convert "/xx/xx/2021/02" to "/xx/xx/year=2021/month=02"
-          val partitionWithName =
-          partitionFragments.zip(partitionSchema).map {
-            case (partition, columnName) =>
-              if (partition.indexOf("=") == -1) {
-                s"${columnName}=$partition"
-              } else {
-                partition
-              }
-          }.mkString("/")
-          val pathWithPartitionName = new Path(basePath, partitionWithName)
-          val partitionValues = parsePartitionValuesFromPath(pathWithPartitionName).toArray
-
-          partitionValues
-        }
-      }
+      val partitionRow = parsePartitionRow(partitionSchema, partitionPath)
       PartitionPath(partitionPath, partitionRow)
     }
   }
 
   /**
-   * Parses partitioning columns' values from the provided path to the partition
+   * Parses partition columns' values from the provided partition's path
+   *
+   * @param partitionColumns partitioning columns identifying the partition
    * @param partitionPath partition's path to parse partitioning columns' values from
    */
-  protected def parsePartitionValuesFromPath(partitionPath: Path): Seq[Any]
+  protected def parsePartitionRow(partitionColumns: Array[String], partitionPath: String): Array[Any]
 
   // TODO eval whether we should just use logger directly
   protected def logWarning(str: => String): Unit
